revived_from_all_failed: optimize code

+++ src/brpc/cluster_recover_policy.h
@@ -62,16 +62,16 @@ public:
     bool StopRecoverIfNecessary();
 
 private:
-    int GetUsableServerCount(int64_t now_ms, const std::vector<ServerId>& server_list);
+    uint64_t GetUsableServerCount(int64_t now_ms, const std::vector<ServerId>& server_list);
 
 private:
     bool _recovering;
     int64_t _minimum_working_instances;
     butil::Mutex _mutex;
-    int64_t _last_usable;
+    uint64_t _last_usable;
     int64_t _last_usable_change_time_ms;
     int64_t _hold_time_ms;
-    int64_t _usable_cache;
+    uint64_t _usable_cache;
     int64_t _usable_cache_time_ms;
 };
